Fix getRequestingUrl() to pass back queue data.
I have been trying to figure out why my bot is broken when queueing tracks. Turns out when getRequestingUrl parses the query but doesn't have any search parameters, it just deletes the original query data that was supposed to me sent to Lavalink causing the most confusing set of issues i've dealt with. This was my solution that works for my bot.